LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/navigation-timing-delivery-type.https_prefetch=false&bypass_cache=true-expected.txt


FAIL PerformanceNavigationTiming.deliveryType test, same origin prefetch. assert_equals: expected (string) "" but got (undefined) undefined
